Cleaned main file. Spelling Corrections.
